God has spoken
