+++ backend/Api/data/SeedData.cs
@@ -1,13 +1,14 @@
 using System;
 using System.Linq;
 using backend.entity.user;
+using backend.entity.utilites;
 using entity.order;
 
 namespace Api.data
 {
     public static class SeedData
     {
-        public static void Initialize(UserContext context)
+        public static void Initialize(DataContext context)
         {
             if (!context.UserSet.Any())
             {
