--- backend/Entity/table/TableContext.cs
@@ -1,22 +1,0 @@
-using entity.order;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-
-namespace backend.entity.table
-{
-    public class TableContext :DataContext
-    {
-        public TableContext (IConfiguration options)
-            : base(options)
-        {
-        }
-
-        public DbSet<Table> Tables { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<Table>().ToTable("Tables");
-        }
-    }
-}
-
