+++ backend/Entity/utilites/Authentication.cs
@@ -1,7 +1,7 @@
 namespace backend.entity.utilites
 {
     public class Authentication
-    {
+    {    
         public string Username { get; set; }
         public string Password { get; set; }
     }
