+++ backend/Entity/utilites/DataContext.cs
@@ -1,24 +1,40 @@
+using backend.entity.product;
+using backend.entity.user;
+using entity.order;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 
 namespace backend.entity.utilites
 {
+
     public class DataContext: DbContext
     {
-        protected readonly IConfiguration Configuration;
+            public DataContext (DbContextOptions<DataContext> options)
+                :base(options)
+            {
+            }
+            public DbSet<User> UserSet { get; set; }
+            public DbSet<Product> ProductSet { get; set; }
+            public DbSet<Option> OptionSet { get; set; }
+            public DbSet<Category> CategorySet { get; set; }
+            public DbSet<ProductConfiguration> ProductConfigurationsSet { get; set; }
+            public DbSet<Table> Tables { get; set; }
+            public DbSet<Order> OrderSet { get; set; }
 
-        public DataContext(IConfiguration configuration)
-        {
-            Configuration = configuration;
+
+            protected override void OnModelCreating(ModelBuilder modelBuilder)
+            {
+                modelBuilder.Entity<Order>().ToTable("Order");
+                modelBuilder.Entity<Table>().ToTable("Tables");
+                modelBuilder.Entity<Product>().ToTable("Product");
+                modelBuilder.Entity<Option>().ToTable("Option");
+                modelBuilder.Entity<Category>().ToTable("Category");
+                modelBuilder.Entity<ProductConfiguration>().ToTable("ProductConfiguration");
+                modelBuilder.Entity<User>()
+                    .ToTable("User")
+                    .HasKey(x => x.Id);
+            }
         }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder options)
-        {
-            // connect to sql server database
-            options
-                .EnableSensitiveDataLogging()
-                .UseSqlServer(@"Server=(localdb)\mssqllocaldb;Database=Test");
-        }
-
     }
-}
+
+
