--- backend/backend/Entity/utilites/Authentication.cs
@@ -1,8 +1,0 @@
-namespace backend.entity
-{
-    public class Authentication
-    {
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
