--- backend/backend/Entity/utilites/DataContext.cs
@@ -1,24 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-
-namespace backend.entity
-{
-    public class DataContext: DbContext
-    {
-        protected readonly IConfiguration Configuration;
-
-        public DataContext(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder options)
-        {
-            // connect to sql server database
-            options
-                .EnableSensitiveDataLogging()
-                .UseSqlServer(@"Server=(localdb)\mssqllocaldb;Database=Test");
-        }
-
-    }
-}
