--- backend/backend/Entity/utilites/Token.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace backend.entity
-{
-    public class Token
-    {
-        public string AccessToken { get; set; }
-        public string RefreshToken { get; set; }
-        public Guid UserId { get; set; }
-    }
-}
