+++ backend/service/WaiterService.cs
@@ -1,15 +1,16 @@
 using System.Collections.Generic;
 using System.Data.SQLite;
+using backend.entity;
 using backend.entity.user;
+using backend.entity.utilites;
 using entity.order;
 
 namespace backend.service
 {
     public class WaiterService
     {
-        private UserContext context;
+        private DataContext context;
       
-        
-     
+
     }
 }
